% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_signature_sample_vector.R
\name{create_signature_sample_vector}
\alias{create_signature_sample_vector}
\title{Create signature sample vector}
\usage{
create_signature_sample_vector(signature, num.mut = 100)
}
\arguments{
\item{signature}{A signature matrix}

\item{num.mut}{Number of mutations generated}
}
\value{
A vector of type character
}
\description{
Simulate sample to match signature distribution (vector of all mutations)
}
\examples{
# Load signature from COSMIC
cosmic.sigs <- get_known_signatures()
cosmic.sig4 <- as.matrix(cosmic.sigs[,4])

# Create vector
create_signature_sample_vector(cosmic.sig4, 100)
}
