% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excerno_vcf.R
\name{excerno_vcf}
\alias{excerno_vcf}
\title{Exercerno VCF}
\usage{
excerno_vcf(
  files,
  artifact = c(),
  method = "nmf",
  num.signatures = 2,
  target.sigs = c()
)
}
\arguments{
\item{files}{VCF files}

\item{artifact}{The signature to consider as an artifact}

\item{method}{A string. The method used to determine the signatures (if not given) and calculate the contributions of each signature}

\item{num.signatures}{Number of signatures. Necessary arugment for when method "linear" is choosen.}

\item{target.sig}{Matrix of the target signatures.Necessary arugment for when method "linear" is choosen.}
}
\value{
Object containing the vcf objects and classification data frame
}
\description{
excerno_vcf() produces filtered vcf files. It uses NMF or nonnegative linear combination of mutation signatures to determine contribution of signatures in samples. Then Bayes' Theroem is used to classify each variant.
}
\examples{

# Load in files
files <- list.files( system.file("extdata", package = "excerno"), pattern = "simulated_sample_sig4..", full.names = TRUE)

exerno_vcf(files)
}
