% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_all_prob.R
\name{extract_all_prob}
\alias{extract_all_prob}
\title{Extract probabilities using bayes' thereom}
\usage{
extract_all_prob(mutation, signatures, contribution)
}
\arguments{
\item{mutation}{Input string of the mutation type to look at. Formatted examples: C[C>A]T, G[T>C]A}

\item{signatures}{Matrix of the signatures relevent in sample}

\item{contributions}{Matrix of how much each signature contributed to the samples}
}
\value{
A vector of the posteriors
}
\description{
extract_all_prob() implements bayes' thereom to get the probabilites of a mutation coming from a specific signature
}
\examples{

cosmic.sigs <- get_known_signatures()

# Get signatures
signatures <- matrix(nrow = 96, ncol = 2)
signatures[,1] <- cosmic.sigs[,4]
signatures[,2] <- get_ffpe_signature()# Naming columns and rows
colnames(signatures) <- c("SBS4", "FFPE")
rownames(signatures) <- get_mutation_types()

# Get contributions
contribution <- matrix(nrow = 2, ncol = 1)
contribution[,1] <- c(0.5, 0.5)
rownames(contribution) <- c("SBS4", "FFPE")

mutation <- "A[C>T]A"
extract_all_prob(mutation, signatures, contribution)
}
