% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_signature_name.R
\name{find_signature_name}
\alias{find_signature_name}
\alias{find_signature_names}
\title{Find signature name}
\usage{
find_signature_name(signature)

find_signature_names(signatures)
}
\arguments{
\item{signature}{A signature matrix}
}
\value{
A string of the most similar signature
}
\description{
Find the cosmic name of the signature that closest resembles a signature
}
\examples{

# Load in a signature
cosmic.sigs <- get_known_signature()
signature.4 <- as.matrix(cosmic.sigs[,4])
rownames(signature.4) <- get_mutation_types()

# Usage
find_signature_name(signature.4)

# Create matrix with multiple signatures
signatures <- cosmic.sigs[,1:3]
find_signature_names(cosmic.sigs[,1:3])
}
