% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_mutation.R
\name{format_mutation}
\alias{format_mutation}
\title{Format mutation}
\usage{
format_mutation(position, sequence, alternate = "")
}
\arguments{
\item{position}{The position of the mutation.}

\item{sequence}{The genomic sequence of type DNAString.}

\item{alternate}{The alternate base of type character. If left empty, format_mutation() will output only the nucleotide base with its surrounding bases.}
}
\value{
A string of the mutation in a certain format
}
\description{
Determine the string format of a position in a genomic sequence
}
\examples{

library(Biostrings)
library(BSgenome.Hsapiens.UCSC.hg38)
 library(stringr)

# Load in the sequence
seq <- getSeq(Hsapiens, "chr1")
format_mutation(20124, seq, "A")
}
