% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_classification.R
\name{get_classification}
\alias{get_classification}
\alias{get_classifications}
\title{Get classification using Bayes' Thereom}
\usage{
get_classification(signatures, contribution)

get_classifications(signatures, contribution)
}
\arguments{
\item{signatures}{A matrix of signatures}

\item{contribution}{A matrix of the contributions coming from each signature}
}
\value{
A data frame with the probabilities for each mutation.
}
\description{
Determine probability of where each mutation came from using Bayes' Thereom.
}
\examples{

# Load in signatures
cosmic.sigs <- get_known_signatures()

# Get signatures
signatures <- matrix(nrow = 96, ncol = 2)
signatures[,1] <- cosmic.sigs[,4]
signatures[,2] <- get_ffpe_signature()
colnames(signatures) <- c("SBS4", "FFPE")
rownames(signatures) <- get_mutation_types()

# Get contributions
contribution <- matrix(nrow = 2, ncol = 1)
contribution[,1] <- c(0.5, 0.5)
rownames(contribution) <- c("SBS4", "FFPE")

# Function usage
get_classification(signatures, contribution)

# For more than one sample
contributions <- matrix(nrow = 2, ncol = 4)
contributions[,] <- sample(0.5, 4, replace = TRUE)
get_classifications(signatures, contributions)
}
