% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mutational_vector.R
\name{get_mutational_vector}
\alias{get_mutational_vector}
\alias{get_mutational_vectors}
\title{Get mutational vector}
\usage{
get_mutational_vector(vcf.file)

get_mutational_vectors(vcf.files)
}
\arguments{
\item{vcf.data}{The name of the vcf file}
}
\value{
A vector of mutation strings
}
\description{
Get the mutational vector from a vcf file
}
\examples{

# Load file for testing
file <- system.file("extdata", "SIMULATED_SAMPLE_SBS4_1.vcf", package = "excerno")

# Load in correct output
cosmic.sigs <- get_known_signatures()
cosmic.sig4 <- as.matrix(cosmic.sigs[, 4])
ffpe.sig <- get_ffpe_signature()

sample.ffpe <- create_signature_sample_vector(ffpe.sig, 500)
sample.sig4 <- create_signature_sample_vector(cosmic.sig4, 500)

test.vector <- c(sample.ffpe, sample.sig4)
vcf.vector <- get_mutational_vector(file)

# Load files for inputing multiple vcfR objects
vcf.files <- list.files(
  system.file("extdata", package = "excerno"),
  pattern = "SIMULATED_SAMPLE_SBS4_\\\\d.vcf", full.names = TRUE)

sample.vectors <- get_mutational_vectors(vcf.files)
}
