% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_cosine_similarity.R
\name{signature_cosine_similarity}
\alias{signature_cosine_similarity}
\title{Signature cosine similarity}
\usage{
signature_cosine_similarity(mutations.vector, signature.matrix)
}
\arguments{
\item{mutations.vector}{A vector of mutations. Type string.}

\item{siganture.matrix}{A signature matrix}
}
\value{
A number between 0 and 1
}
\description{
Calculating the cosine similarity between a vector of mutations and a mutational signature
}
\examples{

library(MutationalPatterns)
library(tidyverse)

cosmic.sig4 <- as.matrix(get_known_signatures()[,4])
sample.sig4 <- create_signature_sample_vector(cosmic.sig4)
signature_cosine_similarity(sample.sig4, cosmic.sig4)
}
