% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_grange_to_vcf.R
\name{write_grange_to_vcf}
\alias{write_grange_to_vcf}
\title{Write a GRange to a vcf file}
\usage{
write_grange_to_vcf(grange, file.name)
}
\arguments{
\item{grange}{A grange object with columns id, ref, alt, qual, filter, info}

\item{file.name}{A string of the new file name. Make sure to end string with ".vcf"}
}
\description{
Write a GRange to a vcf file
}
